test: restructure tests into type-first layout under test/
Move tools/test/{e2e,unit} into test/{e2e,unit,shared} with per-area
subdirectories (installer, updater, core, publish), fix import paths
and REPO_ROOT depths, and update package.json scripts, CI workflow
paths, eslint config (Node globals for test/**) and docs references.

Also switch the unit-test invocation to a glob pattern: newer Node
versions (21+) treat `node --test <dir>` args as globs, so a bare
directory is executed as a test file and fails.
