+++ docs/security.md
@@ -58,7 +58,7 @@ Run through every item; each maps to code that already exists.
   uninit). Exits nonzero on any finding. Vendored `miniz` is excluded.
 - `pnpm test` / `installer/test/test_hash.mjs` — verifies the C hash algorithm matches the JS
   reference, so a tampered package is detected consistently.
-- `tools/test/smoke-security.mjs` — CI security smoke test: boots the installer headless
+- `test/e2e/installer/smoke-security.mjs` — CI security smoke test: boots the installer headless
   (`--smoke-test`) and asserts every state-changing `/api` route rejects a missing/wrong session
   token, valid tokens pass the gate, and no response carries `Access-Control-Allow-Origin`. Runs in
   the CI workflow on Windows after `pnpm upload:local --mode=dev`; run it locally the same way.
