+++ package.json
@@ -7,11 +7,11 @@
     "lint:c": "node tools/format-c.mjs --check",
     "format": "node tools/format-c.mjs --check && prettier --config ./config/prettier.config.js --ignore-path ./config/.prettierignore . --check",
     "format:fix": "node tools/format-c.mjs --write && prettier --config ./config/prettier.config.js --ignore-path ./config/.prettierignore . --write",
-    "test": "node --test \"tools/test/unit/*.test.mjs\"",
+    "test": "node --test \"test/unit/**/*.test.mjs\"",
     "test:hash": "node installer/test/test_hash.mjs",
-    "test:e2e": "node tools/test/e2e/run.mjs",
-    "test:e2e:installer": "node tools/test/e2e/run.mjs --installer",
-    "test:e2e:updater": "node tools/test/e2e/run.mjs --updater",
+    "test:e2e": "node test/e2e/shared/run.mjs",
+    "test:e2e:installer": "node test/e2e/shared/run.mjs --installer",
+    "test:e2e:updater": "node test/e2e/shared/run.mjs --updater",
     "upload": "node --env-file-if-exists=.env tools/publish/upload.mjs",
     "upload:local": "node tools/publish/upload.mjs --local",
     "dev-clean": "node --env-file-if-exists=.env tools/publish/devClean.mjs"
