+++ test/e2e/shared/config.example.mjs
