+++ test/e2e/updater/puppeteer/firefox-updater-test.mjs
