+++ test/unit/publish/config-probe.mjs
@@ -1,4 +1,4 @@
-// tools/test/unit/config-probe.mjs — Print the generated updater-config module
+// test/unit/publish/config-probe.mjs — Print the generated updater-config module
 // for the argv this process was started with (--mode=prod|dev, --local), so the
 // unit test can verify per-mode URLs/flags by spawning child processes.
