+++ test/unit/publish/hashUtils.test.mjs
@@ -1,4 +1,4 @@
-// tools/test/unit/hashUtils.test.mjs — Unit tests for tools/publish/hashUtils.mjs
+// test/unit/publish/hashUtils.test.mjs — Unit tests for tools/publish/hashUtils.mjs
 //
 // hashUtils.mjs imports paths.js, which calls requireMode() at import time, so
 // the test pushes --mode=prod into process.argv before the dynamic import.
